Fixed narrowing casting.

+++ Source/Scanner3DFirmware/Scanner/Scanner.cpp
@@ -75,10 +75,10 @@ Scanner::Points3DData Scanner::Calculate3DPoints(const PointsData& pointsData, f
 
     for (const auto& point : pointsData)
     {
-        const auto radius = static_cast<float>((point.X - s_config.Origin.X) / sin(s_config.CameraLaserInclinationInRad));
-        const auto height = static_cast<float>((s_config.Origin.Y - point.Y) * cos(s_config.AxisCameraInclinationInRad));
+        const auto radius = (point.X - s_config.Origin.X) / sin(s_config.CameraLaserInclinationInRad);
+        const auto height = (s_config.Origin.Y - point.Y) * cos(s_config.AxisCameraInclinationInRad);
 
-        result.emplace_back(Point3D{ radius * sin(trayAngle), height, radius * cos(trayAngle) });
+        result.emplace_back(Point3D{ static_cast<float>(radius * sin(trayAngle)), static_cast<float>(height), static_cast<float>(radius * cos(trayAngle))});
     }
 
     return result;
